Move Article component and assoc. files to Article dir
