+++ src/app/components/Article/index.jsx
@@ -2,9 +2,9 @@ import React, { Fragment, Component } from 'react';
 import Helmet from 'react-helmet';
 import 'isomorphic-fetch';
 import styled from 'styled-components';
-import Header from './Header';
+import Header from '../Header';
 
-import {C_EBON, FF_NEWS_SANS_REG} from '../../lib/constants/styles';
+import {C_EBON, FF_NEWS_SANS_REG} from '../../../lib/constants/styles';
 
 const Headline = styled.h1`
   color: ${C_EBON};
