+++ src/app/components/Article/index.stories.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react'; // eslint-disable-line import/no-extraneous-dependencies
-import Article from './Article';
+import Article from './index';
 
 storiesOf('Article', module).add('default', () => <Article />);
